chest-xray-cancer-detection/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
streamlit>=1.28.0
anthropic>=0.25.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
kaggle>=1.5.16
python-dotenv>=1.0.0
grad-cam>=1.4.8
tqdm>=4.65.0
pytest>=7.4.0
